Implement getTypeDefinitionNode function in the abstract class
