+++ openapi-cli/src/main/java/io/ballerina/openapi/generators/schema/schematypes/AnyDataSchemaType.java
@@ -18,30 +18,14 @@
 
 package io.ballerina.openapi.generators.schema.schematypes;
 
-import io.ballerina.compiler.syntax.tree.AnnotationNode;
-import io.ballerina.compiler.syntax.tree.IdentifierToken;
-import io.ballerina.compiler.syntax.tree.MarkdownDocumentationNode;
-import io.ballerina.compiler.syntax.tree.MetadataNode;
-import io.ballerina.compiler.syntax.tree.Node;
-import io.ballerina.compiler.syntax.tree.TypeDefinitionNode;
 import io.ballerina.compiler.syntax.tree.TypeDescriptorNode;
 import io.ballerina.openapi.exception.BallerinaOpenApiException;
 import io.ballerina.openapi.generators.GeneratorConstants;
 import io.ballerina.openapi.generators.schema.SchemaUtils;
 import io.swagger.v3.oas.models.media.Schema;
 
-import java.util.List;
-
 import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createIdentifierToken;
-import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createNodeList;
-import static io.ballerina.compiler.syntax.tree.NodeFactory.createMarkdownDocumentationNode;
-import static io.ballerina.compiler.syntax.tree.NodeFactory.createMetadataNode;
 import static io.ballerina.compiler.syntax.tree.NodeFactory.createSimpleNameReferenceNode;
-import static io.ballerina.compiler.syntax.tree.NodeFactory.createToken;
-import static io.ballerina.compiler.syntax.tree.NodeFactory.createTypeDefinitionNode;
-import static io.ballerina.compiler.syntax.tree.SyntaxKind.PUBLIC_KEYWORD;
-import static io.ballerina.compiler.syntax.tree.SyntaxKind.SEMICOLON_TOKEN;
-import static io.ballerina.compiler.syntax.tree.SyntaxKind.TYPE_KEYWORD;
 
 /**
  * Generate TypeDefinitionNode and TypeDescriptorNode for schemas without type.
@@ -54,30 +38,6 @@ public class AnyDataSchemaType extends SchemaType {
 
     public AnyDataSchemaType(boolean nullable) {
         this.nullable = nullable;
-    }
-
-    /**
-     * Generate TypeDefinitionNode when no type is given in the schema.
-     * -- ex:
-     * Sample OpenAPI :
-     * <pre>
-     *     Pet:
-     *       description: Details of the pet.
-     *  </pre>
-     * Generated Ballerina type for the schema `Pet` :
-     * <pre>
-     *      public type Pet anydata;
-     * </pre>
-     */
-    @Override
-    public TypeDefinitionNode generateTypeDefinitionNode(Schema<Object> schemaValue, IdentifierToken typeName,
-                                                         List<Node> schemaDoc, List<AnnotationNode> typeAnnotations)
-            throws BallerinaOpenApiException {
-        MarkdownDocumentationNode documentationNode = createMarkdownDocumentationNode(createNodeList(schemaDoc));
-        MetadataNode metadataNode = createMetadataNode(documentationNode, createNodeList(typeAnnotations));
-        return createTypeDefinitionNode(metadataNode, createToken(PUBLIC_KEYWORD), createToken(TYPE_KEYWORD),
-                typeName, this.generateTypeDescriptorNode(schemaValue),
-                createToken(SEMICOLON_TOKEN));
     }
 
     /**
