+++ openapi-cli/src/main/java/io/ballerina/openapi/generators/schema/schematypes/ReferencedSchemaType.java
@@ -18,30 +18,14 @@
 
 package io.ballerina.openapi.generators.schema.schematypes;
 
-import io.ballerina.compiler.syntax.tree.AnnotationNode;
-import io.ballerina.compiler.syntax.tree.IdentifierToken;
-import io.ballerina.compiler.syntax.tree.MarkdownDocumentationNode;
-import io.ballerina.compiler.syntax.tree.MetadataNode;
-import io.ballerina.compiler.syntax.tree.Node;
-import io.ballerina.compiler.syntax.tree.NodeFactory;
-import io.ballerina.compiler.syntax.tree.TypeDefinitionNode;
 import io.ballerina.compiler.syntax.tree.TypeDescriptorNode;
 import io.ballerina.openapi.exception.BallerinaOpenApiException;
 import io.ballerina.openapi.generators.schema.SchemaUtils;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.media.Schema;
 
-import java.util.List;
-
-import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createNodeList;
-import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createToken;
 import static io.ballerina.compiler.syntax.tree.NodeFactory.createIdentifierToken;
-import static io.ballerina.compiler.syntax.tree.NodeFactory.createMarkdownDocumentationNode;
-import static io.ballerina.compiler.syntax.tree.NodeFactory.createMetadataNode;
 import static io.ballerina.compiler.syntax.tree.NodeFactory.createSimpleNameReferenceNode;
-import static io.ballerina.compiler.syntax.tree.SyntaxKind.PUBLIC_KEYWORD;
-import static io.ballerina.compiler.syntax.tree.SyntaxKind.SEMICOLON_TOKEN;
-import static io.ballerina.compiler.syntax.tree.SyntaxKind.TYPE_KEYWORD;
 import static io.ballerina.openapi.generators.GeneratorUtils.extractReferenceType;
 import static io.ballerina.openapi.generators.GeneratorUtils.getValidName;
 
@@ -58,34 +42,6 @@ public class ReferencedSchemaType extends SchemaType {
     public ReferencedSchemaType(OpenAPI openAPI, boolean nullable) {
         this.openAPI = openAPI;
         this.nullable = nullable;
-    }
-
-    /**
-     * Generate TypeDefinitionNode for referenced types.
-     * -- ex:
-     * Sample OpenAPI :
-     * <pre>
-     *      components:
-     *          schemas:
-     *              PetName:
-     *                  type: string
-     *              DogName:
-     *                  $ref: "#/components/schemas/PetName"
-     *  </pre>
-     * Generated Ballerina type for the schema `DogName` :
-     * <pre>
-     *     public type DogName PetName;
-     * </pre>
-     */
-    @Override
-    public TypeDefinitionNode generateTypeDefinitionNode(Schema<Object> schemaValue, IdentifierToken typeName,
-                                                         List<Node> schemaDoc, List<AnnotationNode> typeAnnotations)
-            throws BallerinaOpenApiException {
-        MarkdownDocumentationNode documentationNode = createMarkdownDocumentationNode(createNodeList(schemaDoc));
-        MetadataNode metadataNode = createMetadataNode(documentationNode, createNodeList(typeAnnotations));
-        return NodeFactory.createTypeDefinitionNode(metadataNode, createToken(PUBLIC_KEYWORD),
-                createToken(TYPE_KEYWORD), typeName, this.generateTypeDescriptorNode(schemaValue),
-                createToken(SEMICOLON_TOKEN));
     }
 
     /**
